MB_CUR_MAX has to be of type size_t, as required by POSIX.

+++ src/include/stdlib.h
@@ -104,7 +104,7 @@ typedef __wchar_t wchar_t;
 #endif
 
 // Process wide locale always uses ASCII.
-#define MB_CUR_MAX 1
+#define MB_CUR_MAX ((size_t)1)
 
 // Keep existing code happy that assumes that MB_CUR_MAX_L is a macro.
 #define MB_CUR_MAX_L MB_CUR_MAX_L
